move embeddings stats into bgrepo
